added security to delete_user.php + add_admin.php
added ISSETS to these 2 files so that they can't be accessed by bots

+++ delete_user.php
@@ -1,4 +1,23 @@
-<?php
+<?php    
+ob_start();
+
+include_once 'includes/db_connect.php';
+include_once 'includes/functions.php';
+sec_session_start();
+
+if(isset($_SESSION['admin'])) {
+
+if($_SESSION['admin'] !=2 && $_SESSION['admin'] !=1) {
+    echo '<h2>You are not an Admin!</h2>';     
+    header("refresh:5;url=.index.php?error=1");
+    exit;
+}       
+} else {
+    echo '<h2>Please login first!</h2>';     
+    header("refresh:5;url=index.php?error=1");
+    exit;
+}  
+
 $servername = "localhost";
 $username = "root";
 $password = "root";
@@ -10,10 +29,17 @@ try {
     $conn->setAttribute(PDO::ATTR_ERRMODE, PDO::ERRMODE_EXCEPTION);
 
     // sql to delete a record
-	$id =$_GET['id'];
+	
     
-    echo $id;
+    $admin =$_SESSION['admin'];
+    if($admin != 2) {
+    $id =$_GET['id'];
     $sql = "DELETE FROM members WHERE id=".$id;
+    } else {
+        echo '<h2>You are trying to delete the superadmin!</h2>';     
+        header("refresh:5;url=adminpanel".$admin.".php");
+        exit;
+    }
 
     // use exec() because no results are returned
     $conn->exec($sql);
@@ -26,3 +52,5 @@ catch(PDOException $e)
 
 $conn = null;
 ?>
+
+<?php ob_flush();?>
